move NLPCheckpointConnector
