+++ nemo/collections/nlp/models/nlp_model.py
@@ -15,22 +15,23 @@
 import glob
 import hashlib
 import json
+from nemo.collections.nlp.parts.nlp_overrides import NLPCheckpointConnector
 import os
 import shutil
 import tarfile
 import tempfile
 from typing import Any, Dict, Optional, Union
+from omegaconf import DictConfig, OmegaConf
 
 import torch
-from megatron import mpu
-from megatron.checkpointing import get_checkpoint_version, set_checkpoint_version
-from omegaconf import DictConfig, OmegaConf
+
 from pytorch_lightning import Trainer
 from pytorch_lightning.accelerators.accelerator import Accelerator
-from pytorch_lightning.core.lightning import LightningModule
-from pytorch_lightning.trainer.connectors.checkpoint_connector import CheckpointConnector
-from pytorch_lightning.utilities import rank_zero_only, rank_zero_warn
-from pytorch_lightning.utilities.cloud_io import atomic_save
+from pytorch_lightning.utilities import rank_zero_only
+
+from megatron import mpu
+from megatron.checkpointing import get_checkpoint_version, set_checkpoint_version
+
 from transformers import TRANSFORMERS_CACHE
 
 from nemo.collections.common.tokenizers.huggingface.auto_tokenizer import AutoTokenizer
@@ -474,46 +475,3 @@ class NLPModel(ModelPT, Exportable):
     def output_module(self):
         return self.classifier
 
-
-class NLPCheckpointConnector(CheckpointConnector):
-    """ Override PTL CheckpointConnector to support model parallel checkpoints from Megatron-LM.
-    """
-
-    def __init__(self, trainer):
-        super().__init__(trainer)
-
-    def save_checkpoint(self, filepath, weights_only: bool):
-        """Slightly modified version of PyTorch Lightning's save_checkpoint.
-
-        Args:
-            filepath ([str]): [description]
-            weights_only (bool): [description]
-
-        Returns:
-            [type]: [description]
-        """
-        app_state = AppState()
-        if app_state.model_parallel_size is not None:
-            # filepath needs to be updated to include mp_rank
-            dirname = os.path.dirname(filepath)
-            basename = os.path.basename(filepath)
-            filepath = f'{dirname}/mp_rank_{app_state.model_parallel_rank:02d}/{basename}'
-
-            # dump states as a checkpoint dictionary object
-            checkpoint = self.dump_checkpoint(weights_only)
-
-            # each model parallel rank needs to save a copy of its model
-            if app_state.data_parallel_rank == 0:
-                # write the checkpoint dictionary on the file
-                if self.trainer.accelerator_backend:
-                    checkpoint = self.trainer.accelerator_backend.on_save(checkpoint)
-                try:
-                    atomic_save(checkpoint, filepath)
-                except AttributeError as err:
-                    if LightningModule.CHECKPOINT_HYPER_PARAMS_KEY in checkpoint:
-                        del checkpoint[LightningModule.CHECKPOINT_HYPER_PARAMS_KEY]
-                    rank_zero_warn(
-                        'Warning, `hyper_parameters` dropped from checkpoint.' f' An attribute is not picklable {err}'
-                    )
-                    atomic_save(checkpoint, filepath)
-        return None
